+++ components/clarinet-sdk/browser/src/sdkProxy.ts
@@ -1,196 +1,20 @@
-import { Cl, serializeCVBytes } from "@stacks/transactions";
 import {
   CallFnArgs,
+  ContractOptions,
   DeployContractArgs,
   TransferSTXArgs,
-  ContractOptions,
   type SDK,
-  type TransactionRes,
 } from "@stacks/clarinet-sdk-wasm-browser";
 
-import {
-  parseEvents,
-  serializePostConditions,
-  type CallFn,
-  type CallReadOnlyFn,
-  type DeployContract,
-  type GetDataVar,
-  type GetMapEntry,
-  type MineBlock,
-  type ParsedTransactionResult,
-  type Execute,
-  type TransferSTX,
-  parseCosts,
-} from "../../common/src/sdkProxyHelpers.js";
+import { createSessionProxy, type ProxiedSimnet } from "../../common/src/sdkProxy.js";
 
-/** @deprecated use `simnet.execute(command)` instead */
-type RunSnippet = SDK["runSnippet"];
+export type Simnet = ProxiedSimnet<SDK>;
 
-// because the session is wrapped in a proxy the types need to be hardcoded
-export type Simnet = {
-  [K in keyof SDK]: K extends "callReadOnlyFn"
-    ? CallReadOnlyFn
-    : K extends "callPublicFn" | "callPrivateFn"
-      ? CallFn
-      : K extends "execute"
-        ? Execute
-        : K extends "runSnippet"
-          ? RunSnippet
-          : K extends "deployContract"
-            ? DeployContract
-            : K extends "transferSTX"
-              ? TransferSTX
-              : K extends "mineBlock"
-                ? MineBlock
-                : K extends "getDataVar"
-                  ? GetDataVar
-                  : K extends "getMapEntry"
-                    ? GetMapEntry
-                    : SDK[K];
-};
-
-function parseTxResponse(response: TransactionRes): ParsedTransactionResult {
-  return {
-    result: Cl.deserialize(response.result),
-    events: parseEvents(response.events),
-    costs: parseCosts(response.costs),
-    performance: response.performance,
-  };
-}
-
-export function getSessionProxy() {
-  return {
-    get(session: SDK, prop: keyof SDK, receiver: any) {
-      // some of the WASM methods are proxied here to:
-      // - serialize clarity values input argument
-      // - deserialize output into clarity values
-
-      if (prop === "callReadOnlyFn" || prop === "callPublicFn" || prop === "callPrivateFn") {
-        const callFn: CallFn = (contract, method, args, sender, options) => {
-          const { postConditions, postConditionMode } = serializePostConditions(options);
-          const response = session[prop](
-            new CallFnArgs(
-              contract,
-              method,
-              args.map(serializeCVBytes),
-              sender,
-              postConditions,
-              postConditionMode,
-            ),
-          );
-          return parseTxResponse(response);
-        };
-        return callFn;
-      }
-
-      if (prop === "execute") {
-        const execute: Execute = (snippet) => {
-          const response = session.execute(snippet);
-          return parseTxResponse(response);
-        };
-        return execute;
-      }
-
-      if (prop === "deployContract") {
-        const callDeployContract: DeployContract = (
-          name,
-          content,
-          options,
-          sender,
-          postConditionOptions,
-        ) => {
-          const rustOptions = options
-            ? new ContractOptions(options.clarityVersion)
-            : new ContractOptions();
-          const { postConditions, postConditionMode } =
-            serializePostConditions(postConditionOptions);
-
-          const response = session.deployContract(
-            new DeployContractArgs(
-              name,
-              content,
-              rustOptions,
-              sender,
-              postConditions,
-              postConditionMode,
-            ),
-          );
-          return parseTxResponse(response);
-        };
-        return callDeployContract;
-      }
-
-      if (prop === "transferSTX") {
-        const callTransferSTX: TransferSTX = (amount, recipient, sender, options) => {
-          const { postConditions, postConditionMode } = serializePostConditions(options);
-          const response = session.transferSTX(
-            new TransferSTXArgs(
-              BigInt(amount),
-              recipient,
-              sender,
-              postConditions,
-              postConditionMode,
-            ),
-          );
-          return parseTxResponse(response);
-        };
-        return callTransferSTX;
-      }
-
-      if (prop === "mineBlock") {
-        const callMineBlock: MineBlock = (txs) => {
-          const serializedTxs = txs.map((tx) => {
-            if (tx.callPublicFn) {
-              return {
-                callPublicFn: {
-                  ...tx.callPublicFn,
-                  args_maps: tx.callPublicFn.args.map(serializeCVBytes),
-                },
-              };
-            }
-            if (tx.callPrivateFn) {
-              return {
-                callPrivateFn: {
-                  ...tx.callPrivateFn,
-                  args_maps: tx.callPrivateFn.args.map(serializeCVBytes),
-                },
-              };
-            }
-            if (tx.deployContract) {
-              return {
-                deployContract: {
-                  ...tx.deployContract,
-                },
-              };
-            }
-            return tx;
-          });
-
-          const responses: TransactionRes[] = session.mineBlock(serializedTxs);
-          return responses.map(parseTxResponse);
-        };
-        return callMineBlock;
-      }
-
-      if (prop === "getDataVar") {
-        const getDataVar: GetDataVar = (...args) => {
-          const response = session.getDataVar(...args);
-          const result = Cl.deserialize(response);
-          return result;
-        };
-        return getDataVar;
-      }
-
-      if (prop === "getMapEntry") {
-        const getMapEntry: GetMapEntry = (contract, mapName, mapKey) => {
-          const response = session.getMapEntry(contract, mapName, serializeCVBytes(mapKey));
-          const result = Cl.deserialize(response);
-          return result;
-        };
-        return getMapEntry;
-      }
-
-      return Reflect.get(session, prop, receiver);
-    },
-  };
+export function getSessionProxy(): ProxyHandler<SDK> {
+  return createSessionProxy<SDK, ContractOptions>({
+    CallFnArgs,
+    ContractOptions,
+    DeployContractArgs,
+    TransferSTXArgs,
+  });
 }
